Fixed merge conflicts, minor fixes.

+++ Cryptosystem/src/keyGen/TrapdoorSampler.java
@@ -1,8 +1,10 @@
 package keyGen;
 
+import java.math.BigInteger;
+
 public class TrapdoorSampler {
 
-	public static KeyTuple trapSamp(long modulus, long securityParameter){
+	public static KeyTuple trapSamp(BigInteger modulus, BigInteger securityParameter){
 		
 		long vectorDimentionality = ParamGen.generateDimentionality(modulus, securityParameter);
